requirements.txt
argparse==1.4.0
openpyxl==3.1.2
tensorflow==2.15.1
pandas==2.2.3
memory-profiler==0.61.0
numpy==1.26.4
imageio==2.33.1
imgaug==0.4.0
tqdm==4.66.2
PyYAML==6.0.1
opencv-python==4.8.1.78
scikit-learn==1.4.0
matplotlib==3.8.0
seaborn==0.13.2
tensorflow-addons==0.23.0
vit-keras==0.1.2
keras==2.15.0
nvidia-ml-py3
pywhatkit==5.4
